chore(eth-sender): extact abstract l1 interface and all fee-related code (#2213)
This PR deliberately only moves the code around without refactoring the
logic itself to decrease the risk of error. More PRs will most likely
follow

---------
